refactor: Add 'deno_os' crate
