+++ runtime/lib.rs
@@ -16,6 +16,7 @@ pub use deno_kv;
 pub use deno_napi;
 pub use deno_net;
 pub use deno_node;
+pub use deno_os;
 pub use deno_permissions;
 pub use deno_terminal::colors;
 pub use deno_tls;
@@ -35,7 +36,6 @@ pub mod ops;
 pub mod permissions;
 pub mod signal;
 pub mod snapshot;
-pub mod sys_info;
 pub mod tokio_util;
 pub mod web_worker;
 pub mod worker;
@@ -46,6 +46,7 @@ pub use worker_bootstrap::WorkerExecutionMode;
 pub use worker_bootstrap::WorkerLogLevel;
 
 pub mod shared;
+pub use deno_os::exit;
 pub use shared::runtime;
 
 pub struct UnstableGranularFlag {
@@ -146,12 +147,6 @@ pub static UNSTABLE_GRANULAR_FLAGS: &[UnstableGranularFlag] = &[
     id: 14,
   },
 ];
-
-pub fn exit(code: i32) -> ! {
-  deno_telemetry::flush();
-  #[allow(clippy::disallowed_methods)]
-  std::process::exit(code);
-}
 
 #[cfg(test)]
 mod test {
