+++ runtime/ops/mod.rs
@@ -3,7 +3,6 @@
 pub mod bootstrap;
 pub mod fs_events;
 pub mod http;
-pub mod os;
 pub mod permissions;
 pub mod process;
 pub mod runtime;
