+++ runtime/shared.rs
@@ -42,7 +42,6 @@ extension!(runtime,
     "06_util.js",
     "10_permissions.js",
     "11_workers.js",
-    "30_os.js",
     "40_fs_events.js",
     "40_process.js",
     "40_signals.js",
