+++ tools/core_import_map.json
@@ -239,7 +239,7 @@
     "ext:runtime/06_util.js": "../runtime/js/06_util.js",
     "ext:runtime/10_permissions.js": "../runtime/js/10_permissions.js",
     "ext:runtime/11_workers.js": "../runtime/js/11_workers.js",
-    "ext:runtime/30_os.js": "../runtime/js/30_os.js",
+    "ext:deno_os/30_os.js": "../ext/os/30_os.js",
     "ext:runtime/40_fs_events.js": "../runtime/js/40_fs_events.js",
     "ext:runtime/40_process.js": "../runtime/js/40_process.js",
     "ext:runtime/40_signals.js": "../runtime/js/40_signals.js",
